RDM-12806 refactor CaseLinking packages: part 2
   [RDM-12806](https://tools.hmcts.net/jira/browse/RDM-12806)

Refactor CaseLinking packages; update packages after move.
