+++ src/main/java/uk/gov/hmcts/ccd/data/caselinking/CaseLinkEntity.java
@@ -1,4 +1,4 @@
-package uk.gov.hmcts.ccd.data.caseaccess;
+package uk.gov.hmcts.ccd.data.caselinking;
 
 import javax.persistence.Column;
 import javax.persistence.EmbeddedId;
