+++ src/main/java/uk/gov/hmcts/ccd/endpoint/std/MigrationEndpoint.java
@@ -18,7 +18,7 @@ import uk.gov.hmcts.ccd.auditlog.LogAudit;
 import uk.gov.hmcts.ccd.domain.model.definition.CaseDetails;
 import uk.gov.hmcts.ccd.domain.model.migration.MigrationParameters;
 import uk.gov.hmcts.ccd.domain.model.migration.MigrationResult;
-import uk.gov.hmcts.ccd.domain.service.migration.CaseLinkMigrationService;
+import uk.gov.hmcts.ccd.domain.service.caselinking.CaseLinkMigrationService;
 import uk.gov.hmcts.ccd.domain.service.search.ClassifiedSearchOperation;
 import uk.gov.hmcts.ccd.domain.service.search.SearchOperation;
 import uk.gov.hmcts.ccd.domain.service.search.elasticsearch.ElasticsearchQueryHelper;
