+++ src/main/java/uk/gov/hmcts/ccd/v2/external/controller/TestingSupportController.java
@@ -6,8 +6,8 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
-import uk.gov.hmcts.ccd.domain.model.casedeletion.CaseLinksResource;
-import uk.gov.hmcts.ccd.domain.service.casedeletion.CaseLinkService;
+import uk.gov.hmcts.ccd.domain.model.caselinking.CaseLinksResource;
+import uk.gov.hmcts.ccd.domain.service.caselinking.CaseLinkService;
 import uk.gov.hmcts.ccd.v2.V2;
 
 import javax.inject.Inject;
@@ -17,7 +17,7 @@ import javax.inject.Inject;
 @ConditionalOnProperty(value = "testing.support.endpoint.enabled", havingValue = "true")
 public class TestingSupportController {
 
-    private CaseLinkService caseLinkService;
+    private final CaseLinkService caseLinkService;
 
     @Inject
     public TestingSupportController(CaseLinkService caseLinkService) {
@@ -35,4 +35,5 @@ public class TestingSupportController {
                                     .caseLinks(caseLinkService.findCaseLinks(caseReference))
                                     .build());
     }
+
 }
