+++ src/main/resources/application.properties
@@ -135,11 +135,11 @@ management.health.db.enabled=true
 # Enable disk space health check
 management.health.diskspace.enabled=true
 
-# METRICS ENDPOINT CONFIG
+# TESTING ENDPOINT CONFIG
 # Disable testing support endpoint
 testing.support.endpoint.enabled=${TESTING_SUPPORT_ENABLED:false}
 
-# MIGRATION ENDPOINT CONFIG 
+# MIGRATION ENDPOINT CONFIG
 migrations.endpoint.enabled=${MIGRATIONS_ENDPOINT_ENABLED:false}
 
 # Enable the metrics endpoint
