+++ src/test/java/uk/gov/hmcts/ccd/data/caselinking/CaseLinkRepositoryTest.java
@@ -1,4 +1,4 @@
-package uk.gov.hmcts.ccd.data.caseaccess;
+package uk.gov.hmcts.ccd.data.caselinking;
 
 import org.junit.Before;
 import org.junit.Test;
