+++ src/test/java/uk/gov/hmcts/ccd/domain/service/caselinking/CaseLinkExtractorTest.java
@@ -1,4 +1,4 @@
-package uk.gov.hmcts.ccd.domain.service.casedeletion;
+package uk.gov.hmcts.ccd.domain.service.caselinking;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.core.type.TypeReference;
@@ -16,6 +16,7 @@ import uk.gov.hmcts.ccd.WireMockBaseTest;
 import uk.gov.hmcts.ccd.config.JacksonUtils;
 import uk.gov.hmcts.ccd.domain.model.definition.CaseDetails;
 import uk.gov.hmcts.ccd.domain.model.definition.CaseFieldDefinition;
+import uk.gov.hmcts.ccd.domain.service.common.CaseDataExtractor;
 
 import java.io.IOException;
 import java.util.ArrayList;
@@ -29,7 +30,7 @@ import static org.junit.jupiter.api.Assertions.assertTrue;
 import static org.mockito.ArgumentMatchers.anyMap;
 import static org.mockito.ArgumentMatchers.eq;
 import static org.mockito.Mockito.when;
-import static uk.gov.hmcts.ccd.domain.service.casedeletion.CaseLinkExtractor.TEXT_CASE_REFERENCE;
+import static uk.gov.hmcts.ccd.domain.service.caselinking.CaseLinkExtractor.TEXT_CASE_REFERENCE;
 
 @ExtendWith(MockitoExtension.class)
 class CaseLinkExtractorTest extends CaseLinkTestFixtures {
