+++ src/test/java/uk/gov/hmcts/ccd/endpoint/std/MigrationEndpointTest.java
@@ -8,7 +8,7 @@ import org.mockito.junit.jupiter.MockitoExtension;
 import uk.gov.hmcts.ccd.domain.model.definition.CaseDetails;
 import uk.gov.hmcts.ccd.domain.model.migration.MigrationParameters;
 import uk.gov.hmcts.ccd.domain.model.migration.MigrationResult;
-import uk.gov.hmcts.ccd.domain.service.migration.CaseLinkMigrationService;
+import uk.gov.hmcts.ccd.domain.service.caselinking.CaseLinkMigrationService;
 import uk.gov.hmcts.ccd.domain.service.search.SearchOperation;
 import uk.gov.hmcts.ccd.domain.service.search.elasticsearch.ElasticsearchQueryHelper;
 import uk.gov.hmcts.ccd.endpoint.exceptions.ForbiddenException;
